+++ HISTORY.rst
@@ -1,7 +1,7 @@
 History
 =======
 
-1.6.0 (???)
+1.6.0 (11/04/2018)
 -----------
 
  - Reduce requests required to instantiate applications and users.